New servlet and jsp for mobile device configuration and update so i can update mockey from my crappy-ass blackberry browser. yeah.

+++ src/java/com/mockey/storage/InMemoryMockeyStorage.java
@@ -75,7 +75,7 @@ public class InMemoryMockeyStorage implements IMockeyStorage {
      * </pre>
      */
     public InMemoryMockeyStorage() {
-        this.historyStore.setMaxSize(new Integer(45)); // Was 25...but changed to 45. 
+        this.historyStore.setMaxSize(new Integer(25));  // Careful, more than ~45 and AJAX /JavaScript gets funky.
     }
 
     public Service getServiceById(Long id) {
